requirements.txt
streamlit>=1.39.0
tensorflow-cpu==2.20.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
